dashboard/requirements.txt
streamlit==1.38.0
plotly==5.24.1
requests==2.32.3
python-dotenv==1.0.1
